--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,4 +19,4 @@ members = [
 [profile.release]
 lto = true
 panic = 'abort'
-codegen-units = 1
+# codegen-units = 1