--- a/circuit-construction/src/tests/mod.rs
+++ b/circuit-construction/src/tests/mod.rs
@@ -1,0 +1,1 @@
+mod example_proof;