--- a/hasher/src/tests/hasher.rs
+++ b/hasher/src/tests/hasher.rs
@@ -39,7 +39,7 @@ impl Hashable for TestVector {
 fn test_vectors(test_vector_file: &str, hasher: &mut dyn Hasher<TestVector>) {
     // read test vectors from given file
     let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    path.push("../../oracle/src/tests/test_vectors");
+    path.push("../../../oracle/src/tests/test_vectors");
     path.push(&test_vector_file);
 
     let file = File::open(&path).expect("couldn't open test vector file");