--- a/kimchi/benches/proof_criterion.rs
+++ b/kimchi/benches/proof_criterion.rs
@@ -7,13 +7,13 @@ pub fn bench_proof_creation(c: &mut Criterion) {
 
     let ctx = BenchmarkCtx::new(1 << 10);
     group.bench_function(
-        format!("proof creation (SRS size 2^{}", ctx.srs_size()),
+        format!("proof creation (SRS size 2^{})", ctx.srs_size()),
         |b| b.iter(|| black_box(ctx.create_proof())),
     );
 
     let ctx = BenchmarkCtx::new(1 << 14);
     group.bench_function(
-        format!("proof creation (SRS size 2^{}", ctx.srs_size()),
+        format!("proof creation (SRS size 2^{})", ctx.srs_size()),
         |b| b.iter(|| black_box(ctx.create_proof())),
     );
 
@@ -21,7 +21,7 @@ pub fn bench_proof_creation(c: &mut Criterion) {
 
     group.sample_size(100).sampling_mode(SamplingMode::Auto);
     group.bench_function(
-        format!("proof verification (SRS size 2^{}", ctx.srs_size()),
+        format!("proof verification (SRS size 2^{})", ctx.srs_size()),
         |b| b.iter(|| ctx.batch_verification(black_box(vec![proof.clone()]))),
     );
 }