--- a/oracle/src/tests/mod.rs
+++ b/oracle/src/tests/mod.rs
@@ -1,0 +1,1 @@
+mod poseidon_tests;