--- a/oracle/src/tests/poseidon_tests.rs
+++ b/oracle/src/tests/poseidon_tests.rs
@@ -1,16 +1,15 @@
+use crate::{
+    constants::{PlonkSpongeConstantsKimchi, PlonkSpongeConstantsLegacy},
+    pasta::fp_kimchi as SpongeParametersKimchi,
+    pasta::fp_legacy as SpongeParametersLegacy,
+    poseidon::ArithmeticSponge as Poseidon,
+    poseidon::Sponge as _,
+};
 use mina_curves::pasta::Fp;
 use o1_utils::FieldHelpers;
-use oracle::poseidon::Sponge as _;
 use serde::Deserialize;
 use std::fs::File;
 use std::path::PathBuf; // needed for ::new() sponge
-
-use oracle::poseidon::ArithmeticSponge as Poseidon;
-
-use oracle::constants::PlonkSpongeConstantsKimchi;
-use oracle::constants::PlonkSpongeConstantsLegacy;
-use oracle::pasta::fp_kimchi as SpongeParametersKimchi;
-use oracle::pasta::fp_legacy as SpongeParametersLegacy;
 
 //
 // Helpers for test vectors