--- a/poly-commitment/src/lib.rs
+++ b/poly-commitment/src/lib.rs
@@ -5,4 +5,7 @@ pub mod error;
 pub mod evaluation_proof;
 pub mod srs;
 
+#[cfg(test)]
+mod tests;
+
 pub use commitment::PolyComm;