--- a/poly-commitment/src/tests/batch_15_wires.rs
+++ b/poly-commitment/src/tests/batch_15_wires.rs
@@ -1,13 +1,13 @@
 //! This module tests polynomial commitments, batched openings and
 //! verification of a batch of batched opening proofs of polynomial commitments
 
-use ark_ff::{UniformRand, Zero};
-use ark_poly::{univariate::DensePolynomial, UVPolynomial};
-use colored::Colorize;
-use commitment_dlog::{
+use crate::{
     commitment::{BatchEvaluationProof, CommitmentCurve, Evaluation},
     srs::SRS,
 };
+use ark_ff::{UniformRand, Zero};
+use ark_poly::{univariate::DensePolynomial, UVPolynomial};
+use colored::Colorize;
 use groupmap::GroupMap;
 use mina_curves::pasta::{
     vesta::{Affine, VestaParameters},