--- a/poly-commitment/src/tests/commitment.rs
+++ b/poly-commitment/src/tests/commitment.rs
@@ -1,11 +1,11 @@
-use ark_ff::{UniformRand, Zero};
-use ark_poly::{univariate::DensePolynomial, UVPolynomial};
-use colored::Colorize;
-use commitment_dlog::{
+use crate::{
     commitment::{BatchEvaluationProof, BlindedCommitment, CommitmentCurve, Evaluation, PolyComm},
     evaluation_proof::OpeningProof,
     srs::SRS,
 };
+use ark_ff::{UniformRand, Zero};
+use ark_poly::{univariate::DensePolynomial, UVPolynomial};
+use colored::Colorize;
 use groupmap::GroupMap;
 use mina_curves::pasta::{
     vesta::{Affine, VestaParameters},