--- a/poly-commitment/src/tests/mod.rs
+++ b/poly-commitment/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod batch_15_wires;
+mod commitment;