… oracle/src/tests/test_vectors/kimchi.json
… oracle/src/tests/test_vectors/legacy.json