building and running tests takes ~9min before this PR

(https://github.com/o1-labs/proof-systems/runs/6972612313?check_suite_focus=true)

this PR takes it down to ~7min

It did this by:

1. moving integration tests within their crate (otherwise cargo creates a unique binary for each integration test file)
2. removing `codegen-units = 1` in `Cargo.toml`

(2) means that we get faster compile time but potentially slower code. Since this is a library, it doesn't matter (as the code using this library should set that value anyway) except for the benchmarks. A solution would be to run benchmarks using a different profile setting this back to 1. Is this really a big deal?

after this PR (default `codegen-units`):

```
Proof creation/proof creation (SRS size 2^11                        
                        time:   [417.47 ms 425.38 ms 434.17 ms]

Proof creation/proof creation (SRS size 2^15                        
                        time:   [3.2931 s 3.4628 s 3.6599 s]

Proof creation/proof verification (SRS size 2^15                        
                        time:   [72.719 ms 76.387 ms 80.617 ms]
```

before this PR (`codegen-units = 1`):

```
Proof creation/proof creation (SRS size 2^11                        
                        time:   [406.63 ms 409.72 ms 413.74 ms]
                        change: [-5.7908% -3.6819% -1.6273%] (p = 0.00 < 0.05)
                        Performance has improved.

Proof creation/proof creation (SRS size 2^15                        
                        time:   [3.2260 s 3.2977 s 3.3746 s]
                        change: [-10.213% -4.7694% +0.6106%] (p = 0.14 > 0.05)
                        No change in performance detected.

Proof creation/proof verification (SRS size 2^15                        
                        time:   [65.184 ms 66.319 ms 67.768 ms]
                        change: [-18.028% -13.181% -8.4368%] (p = 0.00 < 0.05)
                        Performance has improved.
```

so negligible change in performance